requirements.txt
# --- App Framework ---
streamlit

# --- Core RAG & LangChain ---
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-openai
langchain-anthropic    # <--- CRITICAL ADDITION for ingest_data_advanced.py

# --- Newer LangChain Integrations ---
langchain-chroma       
langchain-huggingface  

# --- Database & Retrieval ---
chromadb
rank_bm25              
sentence-transformers  

# --- Utilities ---
openai                 
backoff                
python-dotenv
watchdog